# proto/bright/endpoint/endpoint.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "bright/endpoint/endpoint.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Manager"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/create/endpoint": {
      "post": {
        "operationId": "Manager_CreateEndpoint",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/endpointCreateEndpointResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/endpointCreateEndpointRequest"
            }
          }
        ],
        "tags": [
          "Manager"
        ]
      }
    },
    "/delete/endpoint": {
      "post": {
        "operationId": "Manager_DeleteEndpoint",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/endpointDeleteEndpointResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/endpointDeleteEndpointRequest"
            }
          }
        ],
        "tags": [
          "Manager"
        ]
      }
    },
    "/get/endpoint": {
      "post": {
        "operationId": "Manager_GetEndpoint",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/endpointGetEndpointResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/endpointGetEndpointRequest"
            }
          }
        ],
        "tags": [
          "Manager"
        ]
      }
    },
    "/get/endpoints": {
      "post": {
        "operationId": "Manager_GetEndpoints",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/endpointGetEndpointsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/endpointGetEndpointsRequest"
            }
          }
        ],
        "tags": [
          "Manager"
        ]
      }
    }
  },
  "definitions": {
    "basetypeEndpointState": {
      "type": "string",
      "enum": [
        "EndpointDefault",
        "EndpointAvaliable",
        "EndpointParse",
        "EndpointError"
      ],
      "default": "EndpointDefault"
    },
    "brightStringSliceVal": {
      "type": "object",
      "properties": {
        "Op": {
          "type": "string"
        },
        "Value": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "brightStringVal": {
      "type": "object",
      "properties": {
        "Op": {
          "type": "string"
        },
        "Value": {
          "type": "string"
        }
      }
    },
    "brightUint32Val": {
      "type": "object",
      "properties": {
        "Op": {
          "type": "string"
        },
        "Value": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "endpointConds": {
      "type": "object",
      "properties": {
        "ID": {
          "$ref": "#/definitions/brightStringVal"
        },
        "Name": {
          "$ref": "#/definitions/brightStringVal"
        },
        "Address": {
          "$ref": "#/definitions/brightStringVal"
        },
        "State": {
          "$ref": "#/definitions/brightStringVal"
        },
        "RPS": {
          "$ref": "#/definitions/brightUint32Val"
        },
        "IDs": {
          "$ref": "#/definitions/brightStringSliceVal"
        }
      }
    },
    "endpointCreateEndpointRequest": {
      "type": "object",
      "properties": {
        "Info": {
          "$ref": "#/definitions/endpointEndpointReq"
        }
      }
    },
    "endpointCreateEndpointResponse": {
      "type": "object",
      "properties": {
        "Info": {
          "$ref": "#/definitions/endpointEndpoint"
        }
      }
    },
    "endpointDeleteEndpointRequest": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        }
      }
    },
    "endpointDeleteEndpointResponse": {
      "type": "object",
      "properties": {
        "Info": {
          "$ref": "#/definitions/endpointEndpoint"
        }
      }
    },
    "endpointEndpoint": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        },
        "Address": {
          "type": "string"
        },
        "State": {
          "$ref": "#/definitions/basetypeEndpointState"
        },
        "RPS": {
          "type": "integer",
          "format": "int64"
        },
        "Remark": {
          "type": "string"
        },
        "CreatedAt": {
          "type": "string",
          "format": "uint64"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "endpointEndpointReq": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        },
        "Address": {
          "type": "string"
        },
        "State": {
          "$ref": "#/definitions/basetypeEndpointState"
        },
        "RPS": {
          "type": "integer",
          "format": "int64"
        },
        "Remark": {
          "type": "string"
        }
      }
    },
    "endpointGetEndpointRequest": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        }
      }
    },
    "endpointGetEndpointResponse": {
      "type": "object",
      "properties": {
        "Info": {
          "$ref": "#/definitions/endpointEndpoint"
        }
      }
    },
    "endpointGetEndpointsRequest": {
      "type": "object",
      "properties": {
        "Conds": {
          "$ref": "#/definitions/endpointConds"
        },
        "Offset": {
          "type": "integer",
          "format": "int32"
        },
        "Limit": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "endpointGetEndpointsResponse": {
      "type": "object",
      "properties": {
        "Infos": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/endpointEndpoint"
          }
        },
        "Total": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}